fs/ext: extend ext driver and helper scripts for ext4 support
Upgraded the ext filesystem driver in kernel/src/fs/ext/ to support ext4
read and write features:
- Extent tree parsing, block lookup, insertion, merging, depth growth, and
  extent tree deallocation.
- 64-bit group descriptors and flexible block groups (flex_bg) support.
- Extended inodes (256-byte inodes, i_extra_isize, 64-bit file size, and
  extra timestamp fields).
- Directory support for extent-backed directories and fallback for
  indexed directories (EXT4_INDEX_FL).
- Fixed FSCK self-healing pass to correctly map flex_bg metadata blocks
  and extent trees without corrupting on-disk block/inode bitmaps.
- Added safe read_raw_inode deserialization preventing out-of-bounds reads.
- Prevented lock re-entrancy / inversion deadlock by caching gd_size and
  eliminating superblock lock inside write_group_descriptors and get_inode.
- Updated helper scripts in tools/ (format-disk.sh, format-ubuntu-disk.sh,
  run-arch-container.sh) to format images using ext4.
- Comprehensive unit/integration tests added and verified.

+++ kernel/src/fs/ext/dir.rs
@@ -235,7 +235,7 @@ impl ExtInode {
 
         if is_dir {
             let block = self.fs.allocate_block().ok()?;
-            raw_child.i_block[0] = block as u32;
+            raw_child.i_block[0] = block;
             raw_child.i_blocks = self.fs.block_size / 512;
             raw_child.i_size = self.fs.block_size;
